This server can be set up on a local area network by choosing any on computer to be a server node, and using that computer’s private IP address as the server IP address. Once the server is started then the clients can connect to the application with the IP address of the Server Computer
The code uses the concept of sockets and threading.
This will set up a simple Chat Room server and allow multiple clients to connect to it using a client-side script.
Both the server and client script can then be run
   from the Command prompt (in Windows) or from bash
   Terminal (Linux users) by simply typing
python chat_server.py 192.168.55.13 8081
python client.py 192.168.55.13 8081
( you need to specify your own IP address and port number)
The client side script will simply attempt to access the server socket created at the specified IP address and port. Once it connects, it will continuously check as to whether the input comes from the server or from the client, and accordingly redirects output. If the input is from the server, it displays the message on the terminal. If the input is from the user, it sends the message that the users enters to the server for it to be broadcasted to other users.